crawler/requirements.txt
requests>=2.31.0
ddddocr>=1.4.11
pycryptodome>=3.20.0
curl_cffi>=0.13.0
